feat: Configure service CORS to only accept requests from Cloudfront domain
