+++ packages/web/package.json
@@ -29,9 +29,10 @@
     "start": "react-scripts start",
     "build": "react-scripts build",
     "test": "react-scripts test",
-    "e2e": "cypress run --headless",
     "eject": "react-scripts eject",
-    "cypress-open": "cypress open"
+    "e2e": "cypress run --headless",
+    "cypress-open": "cypress open",
+    "deploy": "amplify publish -c -y"
   },
   "eslintConfig": {
     "extends": [
