--- packages/web/src/components/utils/index.ts
@@ -1,1 +1,0 @@
-export * from './string';
