+++ packages/web/src/utils/string.ts
@@ -1,4 +1,15 @@
 import upperFirst from 'lodash/upperFirst';
+import { PokemonNode } from '../types';
+
+export const formatPokemonCaption = (
+  pokemon: Partial<PokemonNode> | undefined,
+) => {
+  if (!pokemon) {
+    return '';
+  }
+  const types = pokemon?.types?.map((type) => type?.name);
+  return `${pokemon?.id} • ${types?.join('/')}`;
+};
 
 /**
  * Formats "example-name" to "Example Name"
